requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asttokens==2.0.5
asgiref==3.4.1
crispy-bootstrap5==0.7.0
dj-database-url==0.5.0
Django==4.0.1
django-crispy-forms==1.13.0
django-filter==21.1
django-multiselectfield==0.1.12
django-storages==1.12.3
gunicorn==20.1.0
jmespath==0.10.0
numpy==1.25.2
pandas==2.0.3
Pillow==9.0.0
psycopg2-binary==2.9.9
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==0.21.1
reportlab==3.6.13
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==5.3.0